Add SFS byte support. Change HTL to byte.
hopsToLive should not be below 1 or above 50, so a byte is enough
storage.

+++ src/freenet/node/fcp/ProbeRequestMessage.java
@@ -63,7 +63,7 @@ public class ProbeRequestMessage extends FCPMessage {
 		try {
 			MHProbe.ProbeType type =  MHProbe.ProbeType.valueOf(fs.get(DMT.TYPE));
 			//If HTL is not defined default to MAX_HTL.
-			final short htl = fs.get(DMT.HTL) == null ? MHProbe.MAX_HTL : fs.getShort(DMT.HTL);
+			final byte htl = fs.get(DMT.HTL) == null ? MHProbe.MAX_HTL : fs.getByte(DMT.HTL);
 			if (htl < 0) throw new MessageInvalidException(ProtocolErrorMessage.INVALID_MESSAGE,
 			                                               "hopsToLive cannot be negative.", null, false);
 			MHProbe.Listener listener = new MHProbe.Listener() {
